Reconcile remapped SDL gamepads

+++ src/fast/backends/gfx_sdl2.cpp
@@ -1,4 +1,4 @@
-#include <cmath>
+#include <math.h>
 #include <stdio.h>
 
 #if defined(ENABLE_OPENGL) || defined(__APPLE__)
@@ -293,7 +293,7 @@ void GfxWindowBackendSDL2::GetActiveWindowRefreshRate(uint32_t* refresh_rate) {
 
     const SDL_DisplayMode* mode = SDL_GetCurrentDisplayMode(display_in_use);
     *refresh_rate =
-        mode != nullptr && mode->refresh_rate > 0.0f ? static_cast<uint32_t>(std::round(mode->refresh_rate)) : 60;
+        mode != nullptr && mode->refresh_rate > 0.0f ? static_cast<uint32_t>(roundf(mode->refresh_rate)) : 60;
 }
 
 static uint64_t previous_time;
